Fixes in RequestedBookServiceImpl.java,ReviewServiceImpl.java, CalculateAverageRatingOnBookImpl.java and RequestedBookStatusException.java Fixes in the method likeRequestedBook(UUID requestedBookId, String userEmail) where likedBy Set was not updated after insert/delete of user Change type of method getAverageRatingOnBook(List<Integer> reviews) from Double to double Removed title from RequestedBookDTO.java Changed parameter name from ISBN to isbn in RequestedBookService.java method getRequestedBookByISBN(String isbn) Changed message of constructor in class RequestedBookStatusException Modified the method filterRequestedBooks to search only by title Removed methods getAllRequestedBooksWithStatus and getTop3FavouriteRequestedBooks Created new class BookStatusTransitionValidator.java

+++ src/main/java/com/kinandcarta/book_library/services/impl/CalculateAverageRatingOnBookImpl.java
@@ -1,26 +1,24 @@
 package com.kinandcarta.book_library.services.impl;
 
-import com.kinandcarta.book_library.entities.Review;
 import com.kinandcarta.book_library.services.CalculateAverageRatingOnBook;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 /**
- * This class is used for implementing calculations used by method in the service logic
- * by the model Review.
+ * Performs calculation to provide a Book object with its average rating.
  */
 @Component
 public class CalculateAverageRatingOnBookImpl implements CalculateAverageRatingOnBook {
 
-    public Double getAverageRatingOnBook(List<Integer> reviews) {
+    public double getAverageRatingOnBook(List<Integer> reviews) {
 
-        int sumReviews = 0;
+        double sumReviews = 0;
 
         for (Integer value : reviews) {
             sumReviews += value;
         }
 
-        return (double) (sumReviews / reviews.size());
+        return sumReviews / reviews.size();
     }
 }
